requirements.txt
aioodbc==0.5.0; python_version >= '3.7'
build==1.2.2.post1; python_version >= '3.8'
certifi==2024.8.30; python_version >= '3.6'
charset-normalizer==3.4.0; python_full_version >= '3.7.0'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
docutils==0.21.2; python_version >= '3.9'
greenlet==3.1.1; python_version >= '3.7'
idna==3.10; python_version >= '3.6'
iniconfig==2.0.0; python_version >= '3.7'
jaraco.classes==3.4.0; python_version >= '3.8'
jaraco.context==6.0.1; python_version >= '3.8'
jaraco.functools==4.1.0; python_version >= '3.8'
keyring==25.5.0; python_version >= '3.8'
markdown-it-py==3.0.0; python_version >= '3.8'
mdurl==0.1.2; python_version >= '3.7'
more-itertools==10.5.0; python_version >= '3.8'
nh3==0.2.19
packaging==24.2; python_version >= '3.8'
pkginfo==1.11.2; python_version >= '3.8'
pluggy==1.5.0; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
pyodbc==5.2.0; python_version >= '3.8'
pyproject-hooks==1.2.0; python_version >= '3.7'
pytest==8.3.3; python_version >= '3.8'
pytest-asyncio==0.24.0; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
pywin32-ctypes==0.2.3; python_version >= '3.6'
readme-renderer==44.0; python_version >= '3.9'
requests==2.32.3; python_version >= '3.8'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rfc3986==2.0.0; python_version >= '3.7'
rich==13.9.4; python_full_version >= '3.8.0'
setuptools==75.6.0; python_version >= '3.9'
sqlalchemy[asyncio]==2.0.36; python_version >= '3.7'
twine==6.0.1; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
urllib3==2.2.3; python_version >= '3.8'
wheel==0.45.1; python_version >= '3.8'